# public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Holodeck Router UI API",
    "description": "REST API for managing VCF 9 Holodeck deployments, Day 2 operations, reservations, commands, user management, and email notifications.",
    "version": "1.2.0"
  },
  "servers": [
    { "url": "/", "description": "Current instance" }
  ],
...
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "holodeck-token",
        "description": "JWT token set as httpOnly cookie after login"
      }
    },
    "schemas": {
...
      "HoloDeckConfig": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "configId": { "type": "string", "description": "ConfigID from holorouter" },
          "description": { "type": "string", "description": "User-friendly description (local metadata)" },
          "notes": { "type": "string", "nullable": true },
          "lastSynced": { "type": "string", "format": "date-time", "nullable": true },
          "vcfVersion": { "type": "string", "description": "From cached config JSON" },
          "instance": { "type": "string", "description": "Running instance ID, if deployed" },
          "targetHost": { "type": "string" },
          "vsanMode": { "type": "string" },
          "depotType": { "type": "string" },
          "dnsDomain": { "type": "string" },
          "hasSiteB": { "type": "boolean" }
        }
      },
...
    }
  },
  "security": [{ "cookieAuth": [] }],
  "paths": {
...
    "/api/holodeck-configs/test-credentials": {
      "post": {
        "tags": ["Holodeck Configs"],
        "summary": "Test ESX/vCenter credentials",
        "description": "Tests ESX/vCenter credentials by running Connect-VIServer from VMware PowerCLI on the holorouter via SSH. Connects and immediately disconnects. labadmin+ only.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["targetHost", "username", "password"],
                "properties": {
                  "targetHost": { "type": "string", "description": "ESX/vCenter hostname or IP" },
                  "username": { "type": "string", "description": "Login username" },
                  "password": { "type": "string", "description": "Login password" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Credential test result",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "message": { "type": "string" }
                  }
                }
              }
            }
          },
          "400": { "description": "Missing or invalid parameters" },
          "403": { "description": "Forbidden — labadmin+ required" }
        }
      }
    },
...
    "/api/holodeck-configs/{id}": {
      "get": {
        "tags": ["Holodeck Configs"],
        "summary": "Get config details",
        "description": "Returns full config details including parsed deployment settings from cached JSON. Superadmins also get the raw JSON.",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" }, "description": "Config ID (from holorouter)" }
        ],
        "responses": {
          "200": {
            "description": "Config details",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "config": { "$ref": "#/components/schemas/HoloDeckConfig" },
                    "deploymentSettings": { "type": "object", "additionalProperties": true },
                    "rawJson": { "type": "object", "description": "Full config JSON (superadmin only)" }
                  }
                }
              }
            }
          },
          "404": { "description": "Config not found" }
        }
      },
      "put": {
        "tags": ["Holodeck Configs"],
        "summary": "Update config metadata",
        "description": "Update description and notes. labadmin+ only.",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "description": { "type": "string" },
                  "notes": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Config updated" },
          "403": { "description": "Forbidden" },
          "404": { "description": "Config not found" }
        }
      },
      "delete": {
        "tags": ["Holodeck Configs"],
        "summary": "Delete config from holorouter",
        "description": "Removes the config JSON file from the holorouter and deletes the local cache entry. Refuses to delete configs with running instances. labadmin+ only.",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" }, "description": "Config ID" }
        ],
        "responses": {
          "200": {
            "description": "Config deleted",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string" }
                  }
                }
              }
            }
          },
          "403": { "description": "Forbidden — labadmin+ required" },
          "404": { "description": "Config not found locally" },
          "409": { "description": "Config has a running instance — remove instance first" }
        }
      }
    },
    "/api/holodeck-configs/{id}/inventory": {
      "post": {
        "tags": ["Holodeck Configs"],
        "summary": "Query VMware inventory from target host",
        "description": "Connects to the config's target vCenter/ESXi host using stored credentials and queries available datastores, networks, clusters, and datacenters via PowerCLI. labadmin+ only.",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" }, "description": "Config ID" }
        ],
        "responses": {
          "200": {
            "description": "VMware inventory",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "datastores": { "type": "array", "items": { "type": "string" }, "description": "Available datastore names" },
                    "networks": { "type": "array", "items": { "type": "string" }, "description": "Available port group / network names" },
                    "clusters": { "type": "array", "items": { "type": "string" }, "description": "Available cluster names (vCenter only)" },
                    "datacenters": { "type": "array", "items": { "type": "string" }, "description": "Available datacenter names (vCenter only)" }
                  }
                }
              }
            }
          },
          "400": { "description": "Config missing credentials or target host" },
          "403": { "description": "Forbidden — labadmin+ required" },
          "404": { "description": "Config not found" },
          "500": { "description": "PowerCLI query failed" }
        }
      }
    },
...
    "/api/reservations/{id}": {
      "delete": {
        "tags": ["Reservations"],
        "summary": "Cancel a reservation",
        "description": "Owner or superadmin can cancel.",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Reservation cancelled" },
          "403": { "description": "Not owner or superadmin" },
          "404": { "description": "Reservation not found" }
        }
      }
    },
...
  }
}